refactor: create auth package #9
