+++ backend/matechai/src/main/java/com/ft/matechai/auth/node/VerificationToken.java
@@ -1,4 +1,4 @@
-package com.ft.matechai.user.node;
+package com.ft.matechai.auth.node;
 
 import lombok.Getter;
 import lombok.Setter;
