+++ backend/matechai/src/main/java/com/ft/matechai/auth/repository/VerificationTokenRepository.java
@@ -1,6 +1,6 @@
-package com.ft.matechai.user.repository;
+package com.ft.matechai.auth.repository;
 
-import com.ft.matechai.user.node.VerificationToken;
+import com.ft.matechai.auth.node.VerificationToken;
 import org.springframework.data.neo4j.repository.Neo4jRepository;
 
 public interface VerificationTokenRepository extends Neo4jRepository<VerificationToken, String> {
